feat: add Kimi Code provider (local CLI OAuth, zero-config) (#7)
Adds Kimi Code provider via localCLI OAuth (reuses local CLI login, zero-config).

Review findings from #7 all addressed in 0b80df8 and verified against real code:
- RFC 3986 strict percent-encoding for refresh form body
- expires_at accepts NSNumber/String, unparsable treated as expired
- two-level ISO8601 parsing; loud throw when usage yields no quotas
- localCLICredential() nil when unconfigured, routing to the /login gear
- booster wallet honors MoneyDTO.currency

+++ Sources/TokenRunwayApp/Ball/LiquidShape.swift
@@ -17,7 +17,7 @@ struct LiquidShape: Shape {
             path.move(to: CGPoint(x: 0, y: surfaceY))
             var x: CGFloat = 0
             while x <= rect.width {
-                let angle = (x / wavelength) * 2 * .pi + phase
+                let angle = Double(x / wavelength) * 2 * .pi + phase
                 let y = surfaceY + amplitude * sin(angle)
                 path.addLine(to: CGPoint(x: x, y: y))
                 x += 2
